add NoItemView

+++ ToDoApp_Youtube/Views/ListView.swift
@@ -12,23 +12,29 @@ struct ListView: View {
     @Environment(ListViewModel.self) var listViewModel
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items){item in
-                ListRowView(item:item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty{
+                NoItemView()
+            }else{
+                List{
+                    ForEach(listViewModel.items){item in
+                        ListRowView(item:item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                    
+                }
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-            
         }
         .listStyle(PlainListStyle())
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
-            leading: EditButton(),
+            leading: EditButton().disabled(listViewModel.items.isEmpty),
             trailing: NavigationLink("Add",destination: AddView()) )
     }
